Add text serialization test
